downtime_monitor/requirements.txt
amqp==5.1.1
asgiref==3.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.5
django-ckeditor==6.5.1
django-cron==0.6.0
django-crontab==0.7.1
django-filter==22.1
django-js-asset==2.0.0
django-model-utils==4.3.1
django-notifications-hq==1.7.0
djangorestframework==3.14.0
idna==3.4
jsonfield==3.1.0
kombu==5.2.4
prompt-toolkit==3.0.36
pytz==2022.7
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
swapper==1.3.0
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
